+++ libcrux-ml-dsa/hax.sh
@@ -8,9 +8,12 @@ function extract_all() {
     
     extract crates/utils/core-models into fstar
 
+    rename_core_models_files crates/utils/core-models
+
     extract crates/utils/intrinsics \
         -C --features simd128,simd256 ";" \
-        into fstar --z3rlimit 80
+        into -i "-core_models::**" \
+        fstar --z3rlimit 80
     
     extract libcrux-ml-dsa \
         -C --features simd128,simd256 ";" \
@@ -64,6 +67,31 @@ function msg() {
     echo -e "$1[$SCRIPT_NAME]$RESET $2"
 }
 
+function rename_core_models_files() {
+    TARGET="$1"
+    shift 1
+    go_to "$TARGET"
+
+    local target_dir="proofs/fstar/extraction"
+    find "$target_dir" -type f -name "Core_models*" | while read -r file; do
+        dir_path="${file%/*}"
+        filename="${file##*/}"
+        new_filename="Libcrux_core_models${filename#Core_models}"
+        mv "$file" "$dir_path/$new_filename"
+    done
+    find "$target_dir" -type f \( -name "*.fst" -o -name "*.fsti" \) -exec sed -i'' \
+        -e 's/module Core_models/module Libcrux_core_models/g' \
+        {} +
+}
+
+function rename_core_models_uses() {
+    local target_dir="proofs/fstar/extraction"
+    find "$target_dir" -type f \( -name "*.fst" -o -name "*.fsti" \) -exec sed -i'' \
+        -e 's/Core_models\.Abstractions/Libcrux_core_models.Abstractions/g' \
+        -e 's/Core_models\.Core_arch/Libcrux_core_models.Core_arch/g' \
+        {} +
+}
+
 function extract() {
     TARGET="$1"
     shift 1
@@ -74,6 +102,7 @@ function extract() {
         msg "$RED" "extract extraction failed for ${BOLD}$1${RESET}"
         exit 1
     }
+    rename_core_models_uses
 }
 
 function help() {