+++ libcrux-psq/src/handshake/dhkem.rs
@@ -13,9 +13,7 @@ use crate::handshake::{
 
 #[derive(TlsSerializeBytes, TlsSize)]
 /// A wrapper around a KEM shared secret.
-///
-/// We don't directly expose this.
-pub(crate) struct DHSharedSecret(Vec<u8>);
+pub struct DHSharedSecret(Vec<u8>);
 
 impl AsRef<[u8]> for DHSharedSecret {
     fn as_ref(&self) -> &[u8] {
@@ -29,6 +27,7 @@ impl AsRef<[u8]> for DHSharedSecret {
     Hash,
     PartialEq,
     Clone,
+    Copy,
     TlsDeserializeBytes,
     TlsSerializeBytes,
     TlsSize,
@@ -36,7 +35,7 @@ impl AsRef<[u8]> for DHSharedSecret {
     TlsDeserialize,
 )]
 /// A wrapper around a KEM public key.
-pub struct DHPublicKey(Vec<u8>);
+pub struct DHPublicKey([u8; 32]);
 
 impl ProvideAuthenticator for DHPublicKey {
     fn authenticator(&self) -> super::types::Authenticator {
@@ -47,6 +46,13 @@ impl ProvideAuthenticator for DHPublicKey {
 impl AsRef<[u8]> for DHPublicKey {
     fn as_ref(&self) -> &[u8] {
         self.0.as_slice()
+    }
+}
+
+/// Import a Diffie-Hellman public key from raw bytes.
+impl DHPublicKey {
+    pub fn from_bytes(value: &[u8; 32]) -> Self {
+        Self(*value)
     }
 }
 
@@ -82,8 +88,26 @@ impl DHPrivateKey {
     pub fn to_public(&self) -> DHPublicKey {
         DHPublicKey(
             secret_to_public(libcrux_ecdh::Algorithm::X25519, &self.0)
+                .expect("secret key is honestly generated X25519 key")
+                .try_into()
                 .expect("secret key is honestly generated X25519 key"),
         )
+    }
+
+    /// Import a Diffie-Hellman private key from raw bytes.
+    pub fn from_bytes(value: &[u8; 32]) -> Result<Self, Error> {
+        // Test whether the key is already clamped to make sure it can't be misused.
+        if !libcrux_ecdh::validate_scalar(libcrux_ecdh::Algorithm::X25519, value).is_ok() {
+            Err(Error::InvalidDHSecret)
+        } else {
+            Ok(Self(Vec::from(value)))
+        }
+    }
+
+    /// Perform a Diffie-Hellman exchange between `self` and `remote_public_key`
+    /// to produce a `DHSharedSecret`
+    pub fn diffie_hellman(&self, remote_public_key: &DHPublicKey) -> Result<DHSharedSecret, Error> {
+        DHSharedSecret::derive(self, remote_public_key)
     }
 }
 
@@ -102,10 +126,26 @@ impl ProvideAuthenticator for DHKeyPair {
 }
 
 impl DHKeyPair {
-    /// Generate a fresh Diffie-Hellman key pair
+    /// Generate a fresh Diffie-Hellman key pair.
     pub fn new(rng: &mut impl CryptoRng) -> Self {
         let sk = DHPrivateKey::new(rng);
         let pk = sk.to_public();
         Self { sk, pk }
+    }
+
+    /// Provide reference to the Diffie-Hellman private key.
+    pub fn sk(&self) -> &DHPrivateKey {
+        &self.sk
+    }
+}
+
+impl From<DHPrivateKey> for DHKeyPair {
+    /// Derive a Diffie-Hellman public key from a private key
+    /// and build a wrapper around both.
+    fn from(sk: DHPrivateKey) -> Self {
+        Self {
+            pk: sk.to_public(),
+            sk,
+        }
     }
 }