+++ libcrux-psq/CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- [#1301](https://github.com/cryspen/libcrux/pull/1301): Fix broken clamping check for imported X25519 secret keys
+- [#1307](https://github.com/cryspen/libcrux/pull/1307): Expose additional functionalities on the DHKEM (https://github.com/jstuczyn)
+- [#1298](https://github.com/cryspen/libcrux/pull/1298): Propagate import/export functions from CMC crate (https://github.com/georgio)
+
+## [0.0.6] (2026-01-22)
+
+- [#1294](https://github.com/cryspen/libcrux/pull/1294): Allow import of X25519 private and public keys (https://github.com/georgio)
 - [#1278](https://github.com/cryspen/libcrux/pull/1278): Allow import of secrets into sessions. (Breaking, since session binding to principal public keys is now optional.)
 - [#1280](https://github.com/cryspen/libcrux/pull/1280): Update dependencies `libcrux-sha3`, `libcrux-ml-kem`, `libcrux-ml-dsa`
 - [#1248](https://github.com/cryspen/libcrux/pull/1248):